+++ utils/FileTimezoneHelper.js
@@ -1,8 +1,8 @@
 const fileTimezone = {
   UTC: 'UTC',
   LOCAL_TIME: 'Site timezone',
-  USE_DEVICE_CONFIG: 'Device configuration'
-  // USE_FILENAME: 'Timezone in filename' // TODO: add this back once we can read timezone from filename
+  USE_DEVICE_CONFIG: 'Device configuration',
+  USE_FILENAME: 'Timezone in filename'
 }
 
 function getTimezoneOptions () {