+++ utils/fileHelper.js
@@ -186,6 +186,15 @@ const archiverDirectory = (sourceDirectory, targetFileName) => {
   archive.finalize()
 }
 
+const getFileTimezoneOffsetFromName = (name) => {
+  // expected format: 20231119T210000+0930_REC_-13.64526+134.30475
+  const pattern = /((?<year>(19|20)[0-9][0-9])[- /.]?(?<month>0[1-9]|1[012])[- /.]?(?<day>0[1-9]|[12][0-9]|3[01]).?(?<hour>[0-1][0-9]|2[0-3])[- :.]?(?<minute>[0-5][0-9])[- :.]?(?<second>[0-5][0-9])?(?<timezone>([+ -])\d*))/ig
+  const regExp = new RegExp(pattern, 'g')
+  var result = regExp.exec(name)
+  if (result === null || (result && result.groups && !result.groups.timezone)) return 0
+  return moment.parseZone(result[0]).utcOffset()
+}
+
 export default {
   getFilesFromDirectoryPath,
   readFile,
@@ -201,6 +210,7 @@ export default {
   getDisplayTimestamp,
   getDisplayFileSize,
   getUtcTimestamp,
+  getFileTimezoneOffsetFromName,
   getFileDuration,
   isSupportedFileExtension,
   isFolder,